+++ test_emu/test_emuloop_run.py
@@ -4,32 +4,8 @@ from mock import patch, MagicMock
 import sys
 inst_mock = MagicMock()
 sys.modules['inst'] = inst_mock
-from emuloop import inclusive_float_range_with_step_flip, DoRun, SetDefinition, magnet_devices
+from emuloop import DoRun, SetDefinition
 import unittest
-
-
-class TestRange(unittest.TestCase):
-
-    def test_GIVEN_float_range_WHEN_get_range_THEN_inclusive_AND_expected_values(self):
-        # GIVEN float range
-        start, stop, step = 0.5, 2, 0.5
-        # WHEN get range
-        returned_range = []
-        for i in inclusive_float_range_with_step_flip(start, stop, step):
-            returned_range.append(i)
-        # THEN inclusive AND expected values
-        self.assertEqual([0.5, 1, 1.5, 2], returned_range)
-
-    def test_GIVEN_inverted_float_range_WHEN_get_range_THEN_inclusive_AND_expected_values(self):
-        # GIVEN float range
-        start, stop, step = 2, 0.5, 0.5
-        # WHEN get range
-        returned_range = []
-        for i in inclusive_float_range_with_step_flip(start, stop, step):
-            returned_range.append(i)
-        # THEN inclusive AND expected values
-        self.assertEqual([2, 1.5, 1, 0.5], returned_range)
-
 
 if six.PY3:
     builtin_eval = 'builtins.eval'
@@ -224,122 +200,3 @@ class TestEmuRunHelpers(unittest.TestCase):
 
     def test_GIVEN_start_stop_not_equal_WHEN_check_set_definition_THEN_definition_is_undefined(self):
         self.assertEqual(self.script_definition.check_set_definition(3.0, 3.1), SetDefinition.SCAN)
-
-
-class TestParameterValidation(unittest.TestCase):
-
-    def setUp(self):
-        self.script_definition = DoRun()
-
-    def test_GIVEN_valid_params_WHEN_validate_THEN_none_returned(self):
-        self.assertIsNone(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="keep", stop_field="keep",
-                                                    step_field="1", custom="None", mevents="10", magnet_device="TF")
-        )
-
-    def test_GIVEN_temp_start_is_keep_and_stop_not_keep_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="keep", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="keep", stop_field="keep",
-                                                    step_field="1", custom="None", mevents="10", magnet_device="TF"),
-            "If start temperature or stop temperature is keep, the other must also be keep\n",
-        )
-
-    def test_GIVEN_temp_stop_is_keep_and_start_not_keep_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="1.0", stop_temperature="keep",
-                                                    step_temperature="1", start_field="keep", stop_field="keep",
-                                                    step_field="1", custom="None", mevents="10", magnet_device="TF"),
-            "If start temperature or stop temperature is keep, the other must also be keep\n",
-        )
-
-    def test_GIVEN_field_start_is_keep_and_stop_not_keep_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_field="keep", stop_field="1.0",
-                                                    step_temperature="1", start_temperature="keep",
-                                                    stop_temperature="keep", step_field="1", custom="None",
-                                                    mevents="10", magnet_device="TF"),
-            "If start field or stop field is keep, the other must also be keep\n",
-        )
-
-    def test_GIVEN_field_stop_is_keep_and_start_not_keep_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_field="1.0", stop_field="keep",
-                                                    step_temperature="1", start_temperature="keep",
-                                                    stop_temperature="keep", step_field="1", custom="None",
-                                                    mevents="10", magnet_device="TF"),
-            "If start field or stop field is keep, the other must also be keep\n",
-        )
-
-    def test_GIVEN_scans_are_defined_AND_steps_are_zero_WHEN_validate_THEN_does_note_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="0", start_field="2", stop_field="30.0",
-                                                    step_field="0", custom="None", mevents="10", magnet_device="TF"),
-            "Cannot step through temperatures when step is zero\n" + "Cannot step through fields when step is zero\n"
-        )
-
-    def test_GIVEN_scans_are_defined_AND_steps_are_less_than_zero_WHEN_validate_THEN_does_note_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="-0.0001", start_field="2", stop_field="30.0",
-                                                    step_field="-2", custom="None", mevents="10", magnet_device="TF"),
-            "Step temperature must be positive\n" + "Step field must be positive\n"
-        )
-
-    def test_GIVEN_try_to_set_zero_field_with_danfysik_and_t20_coils_WHEN_validate_THEN_does_not_validate(self):
-        for invalid_magnet, invalid_magnet_name in {"LF": "Danfysik", "TF": "T20 Coils"}.items():
-            self.assertEqual(
-                self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                        step_temperature="1", start_field="0", stop_field="0",
-                                                        step_field="2", custom="None", mevents="10",
-                                                        magnet_device=invalid_magnet),
-                "Trying to set a zero field without using the active zero field ({}, Active ZF)\n".format(
-                    invalid_magnet_name)
-            )
-
-    def test_GIVEN_try_to_scan_non_zero_field_with_active_zf_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="1", stop_field="20.2",
-                                                    step_field="2", custom="None", mevents="10",
-                                                    magnet_device="ZF"),
-            "Cannot set a non-zero field with the active zero field\n"
-        )
-
-    def test_GIVEN_try_to_set_non_zero_field_with_active_zf_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="1", stop_field="1",
-                                                    step_field="2", custom="None", mevents="10",
-                                                    magnet_device="ZF"),
-            "Cannot set a non-zero field with the active zero field\n"
-        )
-
-    def test_GIVEN_try_to_set_field_using_invalid_magnet_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="1", stop_field="1",
-                                                    step_field="2", custom="None", mevents="10",
-                                                    magnet_device="INVALID"),
-            "Magnet device must be one of {} or N/A\n".format(list(magnet_devices.keys()))
-        )
-
-    def test_GIVEN_try_to_set_field_using_na_magnet_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="1", stop_field="1",
-                                                    step_field="2", custom="None", mevents="10",
-                                                    magnet_device="N/A"),
-            "Field set but magnet devices N/A not in possible devices {}\n".format(list(magnet_devices.keys()))
-        )
-
-    def test_GIVEN_try_to_scan_field_using_na_magnet_WHEN_validate_THEN_does_not_validate(self):
-        self.assertEqual(
-            self.script_definition.parameters_valid(start_temperature="10.5", stop_temperature="1.0",
-                                                    step_temperature="1", start_field="1", stop_field="10.0",
-                                                    step_field="2", custom="None", mevents="10",
-                                                    magnet_device="N/A"),
-            "Field set but magnet devices N/A not in possible devices {}\n".format(list(magnet_devices.keys()))
-        )
